Blog/requirements.txt
crewai==0.11.0
streamlit==1.32.0
google-generativeai==0.3.2
python-dotenv==0.14.0
PyPDF2==3.0.1
fpdf==1.7.2
beautifulsoup4==4.8.0
requests==2.28.2
nltk==3.8.1
langchain-google-genai==0.0.5
reportlab==4.4.1
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
pydantic==2.11.5
typing-extensions==4.14.0
